Fix the failed StorageServiceServerTest test cases
